pyproject.toml
[tool.poetry]
name = "wolt-assignment"
version = "0.1.0"
description = ""
authors = ["pentza <pentti.lahdeaho@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
